// src/forge/schema.gen.ts
export interface components {
    schemas: {
// ...
        ServerResource: {
            id: string;
            /** @enum {string} */
            type: "servers";
            attributes?: {
                id: number;
                credential_id: number | null;
                name: string;
                slug: string;
                /**
                 * @description The type of server.
                 * @example app
                 */
                type: components["schemas"]["ServerType"];
                ubuntu_version: string | null;
                ssh_port: number;
                provider: string;
                identifier: string | null;
                size: string;
                region: string;
                php_version: string | null;
                php_cli_version: string | null;
                opcache_status: string | null;
                database_type: string | null;
                db_status: string | null;
                redis_status: string | null;
                ip_address: string | null;
                private_ip_address: string | null;
                revoked: boolean;
                /**
                 * Format: date-time
                 * @description The date and time the server was created.
                 * @example 2025-07-29T09:00:00Z
                 */
                created_at: string;
                /**
                 * Format: date-time
                 * @description The date and time the server was last updated.
                 * @example 2025-07-30T09:00:00Z
                 */
                updated_at: string;
                connection_status: string | null;
                timezone: string;
                local_public_key: string | null;
                is_ready: boolean;
            };
            relationships?: {
                tags?: {
                    data: components["schemas"]["TagResourceIdentifier"][];
                };
            };
            links: {
                self: components["schemas"]["Link"];
            };
        };
// ...
    };
// ...
}